requirements.txt
Flask==2.2.2
flask-restx==1.0.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
marshmallow==3.16.0
PyJWT==2.4.0
